requirements.txt
Pillow~=9.5.0
amazoncaptcha~=0.5.11
bs4~=0.0.2
datetime~=5.4
openpyxl~=3.1.2
pandas
psycopg2-binary
pytz~=2024.1
selenium~=4.18.1
telebot~=0.0.5
undetected_chromedriver
mysql-connector-python~=8.3.0
python-dotenv~=1.0.1
beautifulsoup4~=4.12.3
requests~=2.30.0
psycopg2~=2.9.9